Common (Windows): allows COM be used in C sources
