+++ CMakeLists.txt
@@ -1033,7 +1033,7 @@ elseif(WIN32)
         src/common/impl/debug_windows.c
         src/common/impl/kmod_windows.c
         src/common/windows/getline.c
-        src/common/windows/com.cpp
+        src/common/windows/com.c
         src/common/windows/registry.c
         src/common/windows/unicode.c
         src/common/windows/variant.cpp
